# ClientSDKs/StoreApi/openapi.g.yaml
openapi: 3.0.1
paths:
  /pet/listPets:
    get:
      summary: List all pets
      operationId: listPets
      x-lz-gencall: PetRepo.ListAsync(callerInfo)
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pet'
        400:
          description: Invalid status value
  /pet/findByStatus:
    get:
      summary: Finds Pets by status
      description: Multiple status values can be provided with comma separated strings
      operationId: findPetsByStatus
      parameters:
      - name: petStatus
        in: query
        description: Status values that need to be considered for filter
        required: true
        explode: true
        schema:
          type: array
          items:
            $ref: '#/components/schemas/PetStatus'
          default: available
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pet'
        400:
          description: Invalid status value
  /pet:
    post:
      summary: Add a new pet to the store
      operationId: addPet
      x-lz-gencall: PetRepo.CreateAsync(callerInfo, body)
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Pet'
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'
        405:
          description: Invalid input
    put:
      summary: Update an existing pet
      operationId: updatePet
      x-lz-gencall: PetRepo.UpdateAsync(callerInfo, body)
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Pet'
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'
        400:
          description: Invalid ID supplied
        404:
          description: Pet not found
        405:
          description: Validation exception
  /pet/findByTags:
    get:
      summary: Finds Pets by tags
      description: Muliple tags can be provided with comma separated strings. Use\ \ tag1, tag2, tag3 for testing.
      operationId: findPetsByTags
      parameters:
      - name: tags
        in: query
        description: Tags to filter by
        required: true
        explode: true
        schema:
          type: array
          items:
            type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pet'
        400:
          description: Invalid tag value
...
  /order:
    post:
      summary: Place an order for a pet
      operationId: placeOrder
      x-lz-gencall: OrderRepo.CreateAsync(callerInfo, body)
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
        description: order placed for purchasing the pet
        required: true
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        400:
          description: Invalid Order
  /blurb:
    post:
      summary: Add a new blurb to the store
      operationId: createBlurb
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Blurb'
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Blurb'
        405:
          description: Invalid input
      x-lz-gencall: BlurbRepo.CreateAsync(callerInfo, body)
    put:
      summary: Update an existing blurb
      operationId: updateBlurb
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Blurb'
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Blurb'
        400:
          description: Invalid ID supplied
        404:
          description: Blurb not found
        405:
          description: Validation exception
      x-lz-gencall: BlurbRepo.UpdateAsync(callerInfo, body)
  /blurb/listBlurbs:
    get:
      summary: List all blurbs
      operationId: listBlurbs
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Blurb'
        400:
          description: Invalid status value
      x-lz-gencall: BlurbRepo.ListAsync(callerInfo)
  /blurb/ListByStatus:
    get:
      summary: List Blurbs by status
      description: Status
      operationId: listBlurbsByStatus
      parameters:
      - name: blurbStatus
        in: query
        required: true
        description: Status value
        schema:
          type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Blurb'
        400:
          description: Invalid status value
      x-lz-gencall: BlurbRepo.ListBlurbsByStatusAsync(callerInfo, blurbStatus)
  /message:
    post:
      summary: Add a new message to the store
      operationId: createMessage
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Message'
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        405:
          description: Invalid input
      x-lz-gencall: MessageRepo.CreateAsync(callerInfo, body)
    put:
      summary: Update an existing message
      operationId: updateMessage
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Message'
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        400:
          description: Invalid ID supplied
        404:
          description: Message not found
        405:
          description: Validation exception
      x-lz-gencall: MessageRepo.UpdateAsync(callerInfo, body)
  /premise:
    post:
      summary: Add a new premise to the store
      operationId: createPremise
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Premise'
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Premise'
        405:
          description: Invalid input
      x-lz-gencall: PremiseRepo.CreateAsync(callerInfo, body)
    put:
      summary: Update an existing premise
      operationId: updatePremise
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Premise'
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Premise'
        400:
          description: Invalid ID supplied
        404:
          description: Premise not found
        405:
          description: Validation exception
      x-lz-gencall: PremiseRepo.UpdateAsync(callerInfo, body)
  /chat:
    post:
      summary: Add a new chat to the store
      operationId: createChat
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Chat'
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chat'
        405:
          description: Invalid input
      x-lz-gencall: ChatRepo.CreateAsync(callerInfo, body)
    put:
      summary: Update an existing chat
      operationId: updateChat
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Chat'
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chat'
        400:
          description: Invalid ID supplied
        404:
          description: Chat not found
        405:
          description: Validation exception
      x-lz-gencall: ChatRepo.UpdateAsync(callerInfo, body)
...
  /public/pet/listPets:
    get:
      summary: List all pets
      operationId: listPets
      x-lz-gencall: PetRepo.ListAsync(callerInfo)
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pet'
        400:
          description: Invalid status value
  /public/pet/findByStatus:
    get:
      summary: Finds Pets by status
      description: Multiple status values can be provided with comma separated strings
      operationId: findPetsByStatus
      parameters:
      - name: petStatus
        in: query
        description: Status values that need to be considered for filter
        required: true
        explode: true
        schema:
          type: array
          items:
            $ref: '#/components/schemas/PetStatus'
          default: available
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pet'
        400:
          description: Invalid status value
  /public/pet/findByTags:
    get:
      summary: Finds Pets by tags
      description: Muliple tags can be provided with comma separated strings. Use\ \ tag1, tag2, tag3 for testing.
      operationId: findPetsByTags
      parameters:
      - name: tags
        in: query
        description: Tags to filter by
        required: true
        explode: true
        schema:
          type: array
          items:
            type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pet'
        400:
          description: Invalid tag value
...
  '/pet/{petId}':
    get:
      summary: Find pet by ID
      description: Returns a single pet
      operationId: getPetById
      x-lz-gencall: PetRepo.ReadAsync(callerInfo, petId)
      parameters:
      - name: petId
        in: path
        description: ID of pet to return
        required: true
        schema:
          type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'
        400:
          description: Invalid ID supplied
        404:
          description: Pet not found
    delete:
      summary: Deletes a pet
      operationId: deletePet
      x-lz-gencall: PetRepo.DeleteAsync(callerInfo, petId)
      parameters:
      - name: petId
        in: path
        description: Pet id to delete
        required: true
        schema:
          type: string
      responses:
        200:
          description: Success
        400:
          description: Invalid ID supplied
        404:
          description: Pet not found
  '/order/{orderId}':
    get:
      summary: Find purchase order by ID
      description: For valid response try integer IDs with value >= 1 and <= 10.\ \ Other values will generated exceptions
      operationId: getOrderById
      x-lz-gencall: OrderRepo.ReadAsync(callerInfo, orderId)
      parameters:
      - name: orderId
        in: path
        description: ID of pet that needs to be fetched
        required: true
        schema:
          type: string
          minimum: 1
          maximum: 10
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        400:
          description: Invalid ID supplied
        404:
          description: Order not found
    delete:
      summary: Delete purchase order by ID
      operationId: deleteOrder
      x-lz-gencall: OrderRepo.DeleteAsync(callerInfo, orderId)
      parameters:
      - name: orderId
        in: path
        description: ID of the order that needs to be deleted
        required: true
        schema:
          type: string
      responses:
        200:
          description: Success
        400:
          description: Invalid ID supplied
        404:
          description: Order not found
...
components:
  schemas:
    Blurb:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
      x-lz-genrepo: true
...
    Pet:
      type: object
      required:
      - name
      - photoUrls
      properties:
        id:
          type: string
        category:
          type: string
        name:
          type: string
          example: Dog
        photoUrls:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        petStatus:
          description: pet status in the store
          default: available
          $ref: '#/components/schemas/PetStatus'
        createUtcTick:
          type: integer
          format: int64
        updateUtcTick:
          type: integer
          format: int64
      x-lz-genrepo: true
    PetStatus:
      type: string
      description: pet status in the store
      enum:
      - available
      - pending
      - sold
    Order:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
          description: Cognito Identity Id of clerk entering order
        petId:
          type: integer
          format: int64
        quantity:
          type: integer
          format: int32
        shipDate:
          type: string
          format: date-time
        orderStatus:
          title: orderStatus
          description: Order Status
          $ref: '#/components/schemas/OrderStatus'
        complete:
          type: boolean
          default: false
        createUtcTick:
          type: integer
          format: int64
        updateUtcTick:
          type: integer
          format: int64
      x-lz-genrepo: true
    OrderStatus:
      type: string
      description: Order Status
      enum:
      - placed
      - approved
      - delivered
...